+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Remove `three` dependency from navcat, move threejs utilities to examples in lieu of a potential @navcat/three package in the future
 - Add `markCylinderArea` compact heightfield function for marking all spans within a cylinder area with a given area ID
+- Change `queryPolygons` to accept a `bounds` Box3 instead of center and halfExtents, same as `queryPolygonsInTile`
 
 ## 0.0.2